+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Declare official support of Python 3.12
 * Cache-Control option when uploading files
 
+### Changed
+* Better help text for --corsRules
+
 ### Infrastructure
 * Remove unsupported PyPy 3.7 from tests matrix and add PyPy 3.10 instead
 * Autocomplete integration tests will now work properly even if tested package has not been installed
@@ -27,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Fast rm sometimes failing due to a rare race condition
 * Fix UnicodeEncodeError in non-Unicode terminals by prioritizing stdout encoding
 * When listing licenses in `license` command only show licenses of `b2` and its dependencies
+* Fix license command failing on Windows when non-UTF8 encoding is the default
 
 ## [3.9.0] - 2023-04-28